Organise the E2E suite by journey rather than by feature
Name each spec for something a person sets out to do, and let the files fall where the flows do. Two seams the old cut had put in the wrong place come apart here: viewing a publication read-only belongs with browsing, not with the route guard it happened to share a file with, and getting data out of the catalogue is a different errand from getting data in.

Validation stops being a file of its own. Nobody sets out to validate — they set out to import, and validation is what pushes back, so it reads as part of that journey.

Undoing an import earns its own spec, apart from deleting a record by hand: they exercise the same compensations but they are not the same errand. The undo that closes an edit stays where it is, being the end of a correction rather than a journey in itself.

No test changed. Every one of the eighteen moved intact, and their titles are identical before and after.
