+++ frontend/e2e/correct-a-publication.spec.ts
